minor: get rid of serialization plugin dependency on the platform
One of the previous commits introduced the usage of the small
jar manifest utility function and therefore introduced the platform
dependency in the serialization plugin, and even special exclusion
in the proguard config.
This doesn't play nice with the embeddable scenarios, especially
for some scripting ones.
This change copies (with a simplification) the utility function to
the use site and removes the dependency.

(cherry picked from commit 5c986e59c8ee7953512cca696c01e3b7bda68e85)