[IR] Don't optimize const floating point operations on JS/WASM
In current implementation, we have several places where the compiler
evaluates constant expressions.
In K1:
 * on the frontend for `const val` property and annotations
 using JVM rules;
 * on the backend in lowering, using the rules of the given platform.

In K2: always on the backend using the rules of the given platform.

The problem with the current approach is that the compiler can't
really evaluate expression following runtime rules because these rules
are not always clear.

After this change, we will ignore constant expressions on floating
point values. On the `const val` property and annotations, the compiler
is going to use JVM rules.

#KT-66922 Fixed

(cherry picked from commit be1c8a3c3866b41b373d065488cb5a368477a300)