FIR DFA: make DataFlowVariable implementations data classes
And don't cache synthetic variables, as they're very lightweight anyway.
Checking whether they are useful can be done by looking at `Flow`: if
there are no implications with that variable in the condition, then
there's no need to add any with it in the effect, either.

This means keeping VariableStorage around is now redundant, as it is
guaranteed to be nothing more than a transient cache. Data flow based
checkers can create new variables if they need to, and it'll work.