Fir2Ir, JVM IR: correctly translate flexible types with raw lower bound
In the newly added test, we have the following type for the single
parameter of `StubSerializer.deserialize`:

    ft<Raw type StubElement<kotlin/Any!>, StubElement<*>?>

This FIR type was incorrectly converted to IR type in
`Fir2IrTypeConverter`: the JVM IR-specific `@RawType` annotation was
added only if the top-level type was raw (i.e. was a ConeRawType), not
if it's a flexible type whose lower bound is raw (i.e. has a RawType
attribute).

Moreover, in case of such type we incorrectly replaced lower bound
arguments (`Any!` in this case, or more generally, upper bounds of the
corresponding type parameters) with the upper bound arguments (`*` in
this case). This led to IR override checker not recognizing that
`KtValueArgumentElementType.deserialize` in the test actually overrides
`StubSerializer.deserialize` because types of parameters
`StubElement<Any>` and `StubElement<*>` are incompatible.

There was also another problem in the backend: during creation of fake
`IrJvmFlexibleType` based on flexibility annotations, we also missed the
fact that lower bound of flexible type can be raw. When this happens, we
need to add the `@RawType` annotation.

 #KT-67142 Fixed

(cherry picked from commit 0a3d31ab754a9841fd7cf2a015e4e5efde2e8836)